256- redirect to rooms if user is already in the room
